### 2017-07-26 da3de4d
Fix #21, update doc and add LICENSE file. fix doc, and cpp code to remove warnings
