+++ setup.py
@@ -9,7 +9,7 @@ from distutils.core import Extension
 
 _MAJOR               = 0
 _MINOR               = 6
-_MICRO               = 3
+_MICRO               = 4
 version              = '%d.%d.%d' % (_MAJOR, _MINOR, _MICRO)
 release              = '%d.%d' % (_MAJOR, _MINOR)
 
