### 2017-07-26 9c5ff8d
fix cpp code
