### 2017-07-27 24c28ea
add MANIFEST and update metadata [ci skip]
