### 2017-07-27 9cb7437
[ci skip] fix license to BSD3
