### 2017-07-27 b6d715c
update changelog and remove duplicated changelog
