### 2017-07-27 27dbcf8
remove duplciated file
