+++ doc/ChangeLog.rst
@@ -1,129 +1,0 @@
-ChangeLog Summary
-#####################
-
-
-Version 0.6.4
----------------
-
-* CHANGES: the bug reported in https://github.com/cokelaer/spectrum/issues/24 is
-  obsolet for the reported module (pburg), which was fixed earlier but the issue
-  was fixed in other module such as psd, parma, correlog
-* add LICENSE file
-* fix warning in cpp code (adding void in func() prototypes)
-
-Version 0.6.3
---------------
-
-* CHANGES: portage nosetests suite to pytest
-* BUG Fixes:
-  * Fix issues https://github.com/cokelaer/spectrum/issues/21 and 
-    https://github.com/cokelaer/spectrum/issues/20 mostly related to
-    compatibility with newest numpy version (1.12)
-
-Version 0.6.1
-------------------
-
-* BUG Fixes:
-  * Issue #5 in pyule sampling not initialised is now fixed
-
-Nov 2013
------------
-* 0.5.7:
-    * doc typo corrected in tukey window
-
-Oct 2012
----------
-* 0.5.6:
-    * refactored documentation using easydev package.
-
-Sept 2012
-----------
-* 0.5.5: 
-    * fix name of the libraries for mac and windows
-    * change setup to manage version properly.
-
-
-March 2012
---------------
-* 0.5.3: add poly2lsf and lsf2poly, add tests, fix bug related to compilation of mydpss.cc
-* 0.5.2: add pmtm
-
-February 2012
---------------
-* 0.5.1: add dpss wtapering windows
-* 0.5.0: 
-   * NPSD replaced by NFFT (qlso not correct for ARMA methods that do not have NFFT since not fourier)
-   * Correlogram replaced by pcorrelogram 
-   * more consistent function and class naming convention 
-   * Update the entire documentation. 
-* 0.4.6: fixed pylab_periodogram, documentation (installation)
-
-January 2012
----------------
-
-* 0.4.5: start to play with Pypi
-October 2011
------------------
-
-* 0.4.4: Start to provide the library on the web www.assembla.com
-
-
-May 2011
-----------
-
-* 0.4.3: :func:`spectrum.periodogram.pdaniell` implemented
-
-April 2011
------------
-
-* 0.4.2: pcovar implemented
-* 0.4.1: pmodcovar implemented
-* 0.4.0: arcovar and modcovar "simplified" version. Documentation updated (tutorial, spectral_estimation, quick start...)
-* 0.3.19: add linear_prediction module with codecs (eg. ac2poly, poly2rc....)
-* 0.3.18 fix bug in levinson (Real data case only) and add ac2poly function.
-* 0.3.17: validation of the modcovar algorithm versus the new arcovar_simplified function.
-* 0.3.16: add a simplified version of arcovar called arcovar_simplified. It is 10 times faster and with a different algorithm provides the same results as arcoar, which validates the two codes!
-* 0.3.15: add corrmtx function. Tested it within music algorithm
-* 0.3.14: cleanup the eigen and music methods by moving the automatic order selection outside the functions.
-* 0.3.13: Add AIC and MDL criteria to deal with automatic eigen values selection in pmusic and pev
-* 0.3.12: test and validate the pmusic and pev pseudo spectrum.
-* 0.3.11: burg and pburg  finalised
-* 0.3.10: tools module cleanup and finalised
-* 0.3.9:  ma fully checked and add pma validated
-* 0.3.8:  minvar fully checked and add pminvar
-* 0.3.7:  aryule fully checked and add pyule
-* 0.3.6:  Speed up by 3 the ARMPSD (renamed to arma2psd)
[…]
