### 2017-08-02 #330 @arfon
@eteq - would you be willing to review this submission for JOSS?
