### 2017-08-02 #348 @whedon
[REVIEW]: Spectrum: Spectral Analysis in Python
**Submitting author:** @cokelaer (<a href="http://orcid.org/0000-0001-6286-1138">Thomas Cokelaer</a>)
**Repository:** <a href="https://github.com/cokelaer/spectrum" target ="_blank">https://github.com/cokelaer/spectrum</a>
**Version:** 0.7.1
