+++ doc/joss/paper.bib
@@ -1,0 +1,41 @@
+
+@book{marple-1987a,
+  address = {Australia, Sydney},
+  author = {Marple, S. L.},
+  publisher = {Prentice Hall},
+  title = {Digital Spectral Analysis With Applications},
+  year = 1987
+}
+
+@book{percival1993spectral,
+  title={Spectral analysis for physical applications: Multitaper and Conventional Univariate Techniques.},
+  author={Percival, Donald B and Walden, Andrew T},
+  year={1993},
+  publisher={cambridge university press}
+}
+
+@article{harris:1978,
+  author = {Harris, F. J.},
+  doi = {10.1109/PROC.1978.10837},
+  issn = {0018-9219},
+  journal = {Proceedings of the IEEE},
+  month = jan,
+  number = 1,
+  pages = {51--83},
+  publisher = {IEEE},
+  title = {On the use of windows for harmonic analysis with the discrete Fouriertransform},
+  url = {http://dx.doi.org/10.1109/PROC.1978.10837},
+  volume = 66,
+  year = 1978
+}
+
+@article{welch,
+  author = {Welch, P.},
+  journal = {IEEE Transactions on Audio and Electroacoustics},
+  volume = 15,
+  number = 2,
+  pages = {70--73},
+  title = {The use of fast Fourier transform for the estimation of power spectra: a method based on time averaging over short, modified periodograms}
+  year = 1967
+  }
+ 
