### 2017-08-02 7d2fdad
Update paper.md
