### 2017-08-03 99c854d
Update paper.bib
