+++ doc/joss/paper.bib
@@ -1,5 +1,5 @@
 
-@book{marple-1987a,
+@book{marple:1987,
   address = {Australia, Sydney},
   author = {Marple, S. L.},
   publisher = {Prentice Hall},
@@ -7,7 +7,7 @@
   year = 1987
 }
 
-@book{percival1993spectral,
+@book{percival:1993,
   title={Spectral analysis for physical applications: Multitaper and Conventional Univariate Techniques.},
   author={Percival, Donald B and Walden, Andrew T},
   year={1993},
@@ -29,7 +29,7 @@
   year = 1978
 }
 
-@article{welch,
+@article{welch:1967,
   author = {Welch, P.},
   journal = {IEEE Transactions on Audio and Electroacoustics},
   volume = 15,
