### 2017-08-03 4df4ad5
Update paper.md
