+++ doc/joss/paper.md
@@ -25,10 +25,11 @@ bibliography: paper.bib
 **Spectrum** is a Python library that includes tools to estimate Power Spectral Densities. Methods
 available are based on Fourier transform, parametric methods or eigenvalues analysis. 
 
-- The Fourier methods are based upon correlogram, periodogram and Welch estimates. Standard tapering windows (Hann, Hamming, Blackman) and more exotic ones are available (DPSS, Taylor, ...).
-- The parametric methods are based on Yule-Walker, BURG, MA and ARMA, covariance and modified covariance methods.
-- Non-parametric methods based on eigen analysis (e.g., MUSIC) and minimum variance analysis are also implemented.
-- Multitapering is also available
+- The Fourier methods are based upon correlogram, periodogram and Welch estimates. Standard tapering windows (Hann, Hamming, Blackman) and more exotic ones are available (DPSS, Taylor, ...) [welch:1967],[marple:1987].
+- The parametric methods are based on Yule-Walker, BURG, MA and ARMA, covariance and modified covariance methods [marple:1987],[percival:1993].
+- Non-parametric methods based on eigen analysis (e.g., MUSIC) and minimum variance analysis are also implemented [marple:1987].
+- Multitapering is also available [percival:1993]
+- Classical tools useful to spectral analysis and more generally signal processing such as window tapering [harris:1978] or transfer function are also available within the library.
 
 -![https://doi.org/10.6084/m9.figshare.5270866.v1](psd_all.png)
 
