### 2017-08-03 #348 @cokelaer
FYI I have added the paper.md and paper.bin in the ./doc/joss directory of the github repository
