### 2017-08-03 6263ac8
Update paper.md
