### 2017-08-03 e33d0be
Update paper.md
