### 2017-08-04 61e5d98
Made matplotlib+pylab dependencies optional (and loading them only in plotting functions)
