+++ src/spectrum/datasets.py
@@ -12,7 +12,6 @@
  
     :Reference: [Marple]_
 """
-from  pylab import plot, linspace, xlabel, ylabel, grid
 import numpy
 from numpy import arange, pi, cos
 
@@ -150,14 +149,9 @@ class TimeSeries():
     def plot(self, **kargs):
         """Plot the data set, using the sampling information to set the x-axis 
         correctly."""
+        from pylab import plot, linspace, xlabel, ylabel, grid
         time = linspace(1*self.dt, self.N*self.dt, self.N)
         plot(time, self.data, **kargs)
         xlabel('Time')
         ylabel('Amplitude')
         grid(True)
-
-
-
-
-
-
