+++ setup.py
@@ -37,7 +37,10 @@ setup(
     package_dir={ '' : 'src' },
 
     # Dependencies
-    install_requires = ['matplotlib', 'numpy', 'scipy', 'easydev'],
+    install_requires=['numpy', 'scipy', 'easydev'],
+    extras_require={
+        'plot': ['matplotlib']
+    },
     data_files = data_files,
     platforms=["Linux"],
     classifiers=["Development Status :: 1 - Planning",
