### 2017-08-07 59e1677
Fix int/float division in the script create_psd
