+++ doc/source/contributions.rst
@@ -1,6 +1,11 @@
 Contributions
 ===============
 
+Contributors
+----------------
+
+Pleas see https://github.com/cokelaer/spectrum/graphs/contributors for details.
+
 How to contribute ?
 -----------------------
 
