### 2017-08-14 cokelaer/spectrum#31 @eteq
Need to include easydev in the docs (or not in the requirements)
The documentation only references numpy, scpiy, and matplotlib as requirements, but `easydev` gets installed when I do a ``pip install``.  It looks to me like `easydev` isn't actually needed at runtime, though?  Perhaps it should be removed from the baseline requirements list (along with sphinx), but have a separate `requirements-dev` list?
