### 2017-08-14 e692ece
Typo fix in install instructions
