### 2017-08-14 cokelaer/spectrum#33 @eteq
Move away from "from spectrum import *" in docs
This is a suggestion which you are free to ignore, but I think is important to think about.

In most contexts, wildcard imports (``from x import *``) is highly-discouraged - even [PEP8](https://www.python.org/dev/peps/pep-0008/) itself says to avoid this.  In science I think this is *especially* important because many users don't have a full understanding of what that's doing and its consequences.

So I would suggest trying to migrate all the examples in the docs to do something like:
```python
import spectrum as sp
p = sp.Periodogram(...)
```
instead of

```python
from spectrum import *
p =Periodogram(...)
```
where feasible.
