### 2017-08-14 988af28
Add travis badge to README
