### 2017-08-14 cokelaer/spectrum#35 @eteq
consider having Periodogram.run be "lazy executed"
While looking over the documentation (and doing some test runs myself) for openjournals/joss-reviews#348, I high-level API/interface item struck me.  Why does ``p.run()`` (or ``p()`` need to happen at all?  All of the subsequent accessor methods or functions (e.g. ``p.plot(...)`` or ``p. pcorrelogram`` or the like) could "lazily" execute ``p.run()`` inside them...? That is, something like:
```python
if not self._has_run:
    self.run()
    self._has_run = True
```

This would free the user from the seemingly unnecessary task of running `run` before doing anything they might actually want to get as an output.
