### 2017-08-14 #348 @eteq
I've completed my review, @arfon and @cokelaer.  Note that in the process I created a few issues in the repository itself (cokelaer/spectrum#31, cokelaer/spectrum#32, cokelaer/spectrum#33, cokelaer/spectrum#34, cokelaer/spectrum#35), but only the first two are really required (they are minor installation/install instructions quirks).

However, there are a number of more general items I found that need attention while working through the checklist.  I've detailed those items below

- [ ] Both the paper and the documentation do not have a clear "statement of need".  That is, they both explain what the package *does*, but not what it's *for*.  I realize that power spectra are quite generally useful across the sciences and engineering, but it's in the JOSS review criteria, and I think it's quite useful to clarify what domain this has been designed for - i.e., the "target audience".
- [ ] Along similar lines: the examples are not "real world".  They are all abstract mathematical exercises, even though power spectra are most useful when applied to real data, whether an engineering signal or a scientific time series.  While not strictly mandatory, an example or two that either uses or simulates a "real world" case would improve the documentation a lot.
- [ ] There are no clear community guidelines.  The contribution section simply says to look at the github cite, but says nothing about whether one should raise issues, create PRs, what sort of credit this leads to, etc.  (This rolls into the next bullet point for probably obvious reasons)  A github-style ``CONTRIBUTING.md`` (or rst) document would be an easy way to resolve this.
- [ ] There is only one author of the paper (@cokelaer), but the commit logs clearly show ~5 other contributors.  If I understand right, these should also be coauthors of the JOSS paper (@arfon, is that correct?).  And even if not, they should be referenced somewhere in the docs as contributors - currently there's a "by reference" that simply points to the github contributors interface, which is not necessarily a reliable measure (some contributors might not have github accounts, etc).
- [ ] While the various packages are a bit hazy on citation expectations, it seems like this work should cite numpy, matplotlib, and possibly scipy given that all are requirements and critical to the functioning of this package.  Some guidelines for that are [here](https://www.scipy.org/citing.html)

And one final item that doesn't necessarily require action, but might be worth considering now:
- [ ] I found `spectrum` to be a rather confusing name.  It has a lot of overloaded meanings in a lot of different scientific contexts. I realize this ship has probably sailed, but I thought I'd at least bring it up: have you (@cokelaer) considered using a more specific name like "power_spectrum"?  I know plenty of people who wouldn't be able to use this solely because they have their own local package named `spectrum` and it would lead to a package namespace conflict...

I think the above puts this at "needs minor revision" (or at least arguments/explanations for why revision is not required).
