### 2017-08-15 ca698aa
Fixing issue #31 to have a clean requirements + requirements-dev file
