+++ doc/source/install.rst
@@ -1,12 +1,12 @@
-Installation and source 
+Installation and source
 =========================
 
 **Spectrum** is available on `PYPi <http://pypi.python.org/pypi>`_, so you should be able to type::
 
-    pip install spectrum 
+    pip install spectrum
 
 
-Since spectrum depends on other python packages such as Numpy, Matplotlib and Scipy they will be installed automatically (if not already installed). 
+Since **spectrum** depends on other python packages such as Numpy, Matplotlib and Scipy they will be installed automatically (if not already installed).
 
 You can also install them yourself by typing::
 
@@ -15,24 +15,29 @@ You can also install them yourself by typing::
 **Spectrum** source code is available on Github https://github.com/cokelaer/spectrum
 
 
-
-
 Conda installation
 ========================
 
-Spectrum is not available on CONDA. However, you can install dependencies as
-follows::
+**Spectrum** is now available on CONDA. For Linux and MAC users, if you prefer to use conda, please use::
 
-    conda install numpy matplotlib scipy 
-
-I put a todo issue on the github account if you want to do the full portage on
-conda-forge for instance (https://github.com/cokelaer/spectrum/issues/28)
-
-For Linux and MAC users, if you prefer to use conda (avoiding 
-compilation of dependencies), please use::
-
-    conda config --add channels conda-forge 
+    conda config --add channels conda-forge
     conda install spectrum
+
+
+Note for developers
+========================
+
+Developpers who want to get the source code can clone the repository::
+
+    git clone git@github.com:cokelaer/spectrum.git
+    cd spectrum
+    python setup.py install
+
+
+Then, you can test the library using **pytest** or compile the documentation
+with Sphinx. To do so, install sphinx and other dependencies::
+
+    pip install --file requirements-dev.txt
 
 
 
