+++ requirements.txt
@@ -1,5 +1,3 @@
-easydev
 numpy
 scipy
 matplotlib
-sphinx
