+++ setup.cfg
@@ -1,14 +1,3 @@
-[nosetests]
-where=test
-#with_coverage=
-#cover_package=spectrum
-#cover_erase=
-verbosity=2
-#cover_html=
-#cover_html_dir=coverage
-
-#[develop]
-#install-dir=../lib/python2.7/site-packages
 
 [upload_docs]
 upload_dir=doc/build/html/
@@ -16,4 +5,11 @@ upload_dir=doc/build/html/
 
 [upload_sphinx]
 upload-dir = doc/build/html
+
+[aliases]
+test=pytest
+
+[tool:pytest]
+# do not use --cov because it interfers with travis command
+addopts= --durations=10 --verbose  -n 1 --cov spectrum --cov-report term-missing
 
