### 2017-08-15 #348 @arfon
Many thanks for your thorough review @eteq!

> If I understand right, these should also be coauthors of the JOSS paper (@arfon, is that correct?)

We don't police authorship on JOSS submissions but yes, I think it would be appropriate to acknowledge the other contributors, perhaps by linking to the GitHub contribution graph?
 https://github.com/cokelaer/spectrum/graphs/contributors
