### 2017-08-15 #348 @eteq
Thanks for the clarification, @arfon - actually this one *does* link to the graph from the docs/README.  But probably it should also be linked in the paper then?

In addition, I'd encourage @cokelaer to consider what `spectrum`'s policy on authorship *should* be when addressing the contributor/community guidelines. I don't know if this has happened here, but in some cases the contributor graph doesn't properly catch everyone (e.g. commits with github-unrecognized email addresses).  But as long as there's a clear understanding, all is fine.

> We don't police authorship on JOSS submissions

Perhaps a discuss for elsewhere, or an already-had discussion (feel free to point me where if so, @arfon), but I wonder if there *should* be a policy on this?
