### 2017-08-16 #348 @danielskatz
@eteq - there has been a lot of research about authorship and contributorship in general, and some for software specifically.

One project worth noting is [Project CRediT](http://docs.casrai.org/CRediT) which arose out of trying to understand and then standardize contributor roles in biomedical papers.  Of the 14 different [possible roles](http://dictionary.casrai.org/Contributor_Roles) that it found, some include writing text, but many do not.  Similarly, in a software project, I think there are roles that lead to commits, and other roles that do not.  And some committers, e.g. an administrator who adds a license file upon instruction from the project lead, may not really be contributors to the project.

Since the set of contributors is not the same as the set of committers, I don't believe that JOSS should make any general assumptions about who the contributors are based on who the committers are, but rather, the submitting author should tell us who the contributors are, and the editor and reviewer should point out any problems they see, such as you have brought up here.  Maybe this is a new point we need to add to the reviewer checklist, as part of the authorship item.  I'll create a PR for this.
