### 2017-08-16 #348 @eteq
Thanks for the links, @danielskatz - I know some of that work but not that one.

I agree that it's not JOSS's role to set specific authorship rules, but it *might* be sensible to have some guidelines. This is particularly relevant because the review for software like this can be cross-disciplinary, where it's not clear what community's "rules" apply.  For example, in my science domain I'd say it *is* inappropriate to not include code contributors.  But I don't know for sure that the author of this package is in my domain.  As you can see above, this lead me to phrase this as "consider what the right policy is", but it would be useful if JOSS had some guidelines about what I should do in this sort of situation so I'm not making this up as I go along :wink:.
