### 2017-08-16 #348 @cokelaer
@eteq @arfon thanks alot for the review. This will help improving the software indeed. I won't go through all the reviews right now but will come back to you very soon. Just a quick comment on the contributors. I am aware of the contributions and had put an issue to remind myself to add a list of contributors I let you know once I have an updated version.
