### 2017-08-16 #348 @danielskatz
@eteq - My opinion, for what it's worth, is that asking reviewers to check on this is probably the best we can do for now, to ensure that other reviewers are as conscientious as you have been, and if there's a question they have, they bring it up.
