### 2017-08-17 27302c1
Flip PSD for the real-data case
