+++ test/pytest.ini
@@ -1,0 +1,2 @@
+[pytest]
+addopts = "--cov-report=term-missing"
