+++ test/test_window.py
@@ -4,26 +4,33 @@ import pytest
 
 
 
-class test_class_Window():
-    def __init__(self):
-        self.w = Window(65, name='hann')
-    def test_Window(self):
-        self.w.enbw
-        self.w.mean_square
-        self.w.frequencies
-        self.w.plot_time_freq()
-    def test_nfft(self):
-        self.w.compute_response(NFFT=32)
 
-    def test_error(self):
-        try:
-            w = Window(64, name='wrong')
-            assert False
-        except:
-            assert True
-    def test_info(self):
-        self.w.info()
-        print(self.w)
+def test_class_Window():
+    w = Window(65, name='hann')
+    w.enbw
+    w.mean_square
+    w.frequencies
+    w.plot_time_freq()
+    w.compute_response(NFFT=32)
+
+    try:
+        w = Window(64, name='wrong')
+        assert False
+    except:
+        assert True
+    w.info()
+    print(w)
+
+    # recompute response
+    w = Window(65, name='hann')
+    w.response
+    w.plot_frequencies(maxdB=120, mindB=-100)
+
+    try:
+        w = Window(65, name="dummy")
+        assert False
+    except:
+        assert True
 
 
 #unittest of create_window
@@ -33,6 +40,7 @@ def test_create_window_error():
         assert False
     except:
         assert True
+
 
 #test that create_window(N, name) works for all valid names
 @pytest.mark.parametrize('test_window_name,length', 
@@ -53,12 +61,30 @@ def test_create_window(test_window_name, length):
       ('poisson',      {'alpha': 2}),
       ('poisson_hanning', {'alpha': 2}),
       ('cauchy',   {'alpha': 3})])
+
+
 def test_check_window_switch(name, param):
     f = eval('window_'+name)
     w1 = f(64, **param)
     w2 = create_window(64, name, **param)
     for x,y in zip(w1,w2):
         assert x==y
+
+def test_create_window_others():
+    try:
+        create_window(11, "hamming", dummy=1)
+        assert False
+    except ValueError:
+        assert True
+
+    try:
+        create_window(11, "kaiser", beta=0.5)
+        create_window(11, "kaiser", dummy=1)
+        assert False
+    except ValueError:
+        assert True
+
+
 
 def test_bartlett():
     """unit and functional test window_bartlett"""
@@ -82,6 +108,8 @@ def test_kaiser():
     window_kaiser(8, method='other')
     window_kaiser(8, method='numpy')
 
+    assert window_kaiser(1, method='numpy') == np.array([1])
+
 
 def test_blackman():
     """unit and functional test window_bartlett"""
@@ -93,6 +121,8 @@ def test_blackman():
         assert_almost_equal(x, y)
     for x, y in zip(window_blackman(7, alpha=0.16), vec7):
         assert_almost_equal(x, y)
+
+    assert window_blackman(1, alpha=0.16) == np.array([1])
 
 
 def test_hann():
@@ -138,20 +168,29 @@ def test_gaussian():
 def test_cauchy():
     window_cauchy(64)
 
+
 def test_cosine():
     window_cosine(64)
+    assert window_cosine(1) == np.array([1.])
+
 
 def test_riemann():
     window_riemann(64)
 
+
 def test_lanczos():
     window_lanczos(64)
+    assert window_lanczos(1) == np.array([1.])
+
 
 def test_poisson():
     window_poisson(64)
+    #assert window_poisson(1) == np.array([1.])
+
 
[…]
