### 2017-08-19 bb34a68
Fix documentation and examples
