### 2017-08-19 7d54fd8
Fix doc/examples/pep8
