### 2017-08-19 cee9eec
Fix documentation/pep8/examples
