+++ src/spectrum/covar.py
@@ -1,21 +1,20 @@
 import numpy
 from .psd import ParametricSpectrum
-debug = False
 
 
 __all__ = ["arcovar", "arcovar_marple", "pcovar", 'arcovar_marple']
 
-def arcovar_marple(x, order):
+def arcovar_marple(x, order, debug=False):
     r"""Estimate AR model parameters using covariance method
-    
-    This implementation is based on [Marple]_. This code is far more 
+
+    This implementation is based on [Marple]_. This code is far more
     complicated and slower than :func:`arcovar` function, which is now the official version.
     See :func:`arcovar` for a detailed description of Covariance method.
-    
-    This function should be used in place of arcovar only if order<=4, for 
-    which :func:`arcovar` does not work. 
 
-    Fast algorithm for the solution of the covariance least squares normal 
+    This function should be used in place of arcovar only if order<=4, for
+    which :func:`arcovar` does not work.
+
+    Fast algorithm for the solution of the covariance least squares normal
     equations from Marple.
 
     :param array X:  Array of complex data samples
@@ -29,51 +28,51 @@ def arcovar_marple(x, order):
         * PV   - store linear prediction coefficients
 
     .. note:: this code and the original code in Marple diverge for ip>10.
-        it seems that this is related to single precision used with 
-        complex type in fortran whereas numpy uses double precision for 
+        it seems that this is related to single precision used with
+        complex type in fortran whereas numpy uses double precision for
         complex type.
 
-    :validation: the AR parameters are the same as those returned by 
+    :validation: the AR parameters are the same as those returned by
         a completely different function :func:`arcovar`.
-        
+
     :References: [Marple]_
     """
     assert len(x) >= order, "X must be dimensioned >=N"
-    
+
     #   ----------------------------------------------------- Initialization
     x = numpy.array(x)
     N = len(x)
-    
-    
+
+
     # Equations 8.C.42
     r0 = sum(abs(x)**2.)
     r1 = abs(x[0])**2
     rN = abs(x[N-1])**2
-    
+
     pf = r0 - r1
     pb = r0 - rN
-    
+
     delta = 1. - r1 / r0
-    gamma = 1. - rN / r0 
+    gamma = 1. - rN / r0
 
     c = numpy.zeros(N, dtype=complex)
     d = numpy.zeros(N, dtype=complex)
     r = numpy.zeros(N, dtype=complex)
     af = numpy.zeros(N, dtype=complex)
     ab = numpy.zeros(N, dtype=complex)
-    
+
     c[0] = x[N-1].conjugate() / r0
     d[0] = x[0].conjugate() / r0
-    
+
     # special case
     if order == 0:
         pf = r0 / float(N)
         pb = pf
         return af, pf, ab, pb, 0
-    
+
     # ----------------------------------------------------------  MAIN LOOP
 
-    #ip +1 because we want to enter in the loop to run the first part of the code. 
+    #ip +1 because we want to enter in the loop to run the first part of the code.
     pbv = []
     for m in range(0, order+1):
         if debug:
@@ -138,7 +137,7 @@ def arcovar_marple(x, order):
                 save = af[k]
                 af[k] = save + c1 * ab[m-k-1] # Eq. (8.C.18)
                 ab[m-k-1] = ab[m-k-1] + c2 * save   # Eq. (8.C.21)
-                
+
                 save = c[k]
                 c[k] = save + c3*d[k]       # Eq. (8.C.37)
                 d[k] = d[k] + c4*save       # Eq. (8.C.38)
@@ -174,7 +173,7 @@ def arcovar_marple(x, order):
             print('pf and pb', pf, pb)
         if pf > 0 and pb > 0:
[…]
