### 2017-08-19 ad89ab9
Fix bug in AKICc
