### 2017-08-19 ef507f0
add a toy example for the real case (two sinuses + noise)
