+++ src/spectrum/datasets.py
@@ -4,100 +4,98 @@
 
     .. autosummary::
 
-        data_cosine 
+        data_cosine
         marple_data
-        TimeSeries 
-        
+        TimeSeries
+
     .. codeauthor:: Thomas Cokelaer 2011
- 
+
     :Reference: [Marple]_
 """
 import numpy
-from numpy import arange, pi, cos
-
-
+from numpy import arange, pi, cos, sin
+from numpy.random import randn
 
 
 #example marple app 3.A
 A_cholesky = numpy.matrix([[2+0.j, .5-0.5j,-.2+.1j],
                            [.5+.5j,1,.3-0.2j],
                            [-.2-.1j,.3+.2j,.5]], dtype=complex)
-a_cholesky = numpy.array([2+0.j, .5-0.5j, 1., -.2+.1j,.3-0.2j,.5], 
+a_cholesky = numpy.array([2+0.j, .5-0.5j, 1., -.2+.1j,.3-0.2j,.5],
                          dtype=complex)
 B_cholesky = numpy.array([1+3j,2-1j,.5+.8j], dtype=complex)
-#should return 
-sol_cholesky = numpy.array([ 0.95945946+5.25675676j, 
-                            4.41891892-7.04054054j, 
+#should return
+sol_cholesky = numpy.array([ 0.95945946+5.25675676j,
+                            4.41891892-7.04054054j,
                             -5.13513514+6.35135135j])
 
 
 #: 64-complex data length from Marple reference [Marple]_
 marple_data = [
   1.349839091+  2.011167288j,
- -2.117270231+  0.817693591j, 
- -1.786421657-  1.291698933j, 
-  1.162236333-  1.482598066j, 
-  1.641072035+  0.372950256j, 
-  0.072213709+  1.828492761j, 
- -1.564284801+  0.824533045j, 
- -1.080565453-  1.869776845j, 
-  0.927129090-  1.743406534j, 
-  1.891979456+  0.972347319j, 
- -0.105391249+  1.602209687j, 
- -1.618367076+  0.637513280j, 
- -0.945704579-  1.079569221j, 
-  1.135566235-  1.692269921j, 
-  1.855816245+  0.986030221j, 
- -1.032083511+  1.414613724j, 
- -1.571600199+  0.089229003j, 
- -0.243143231-  1.444692016j, 
-  0.838980973-  0.985756695j, 
-  1.516003132+  0.928058863j, 
-  0.257979959+  1.170676708j, 
- -2.057927608+  0.343388647j, 
- -0.578682184-  1.441192508j, 
-  1.584011555-  1.011150956j, 
-  0.614114344+  1.508176208j, 
- -0.710567117+  1.130144477j, 
- -1.100205779-  0.584209621j, 
-  0.150702029-  1.217450142j, 
-  0.748856127-  0.804411888j, 
-  0.795235813+  1.114466429j, 
- -0.071512341+  1.017092347j, 
- -1.732939839-  0.283070654j, 
-  0.404945314-  0.781708360j, 
-  1.293794155-  0.352723092j, 
- -0.119905084+  0.905150294j, 
- -0.522588372+  0.437393665j, 
- -0.974838495-  0.670074046j, 
-  0.275279552-  0.509659231j, 
-  0.854210198-  0.008278057j, 
-  0.289598197+  0.506233990j, 
- -0.283553183+  0.250371397j, 
- -0.359602571-  0.135261074j, 
-  0.102775671-  0.466086507j, 
- -0.009722650+  0.030377999j, 
-  0.185930878+  0.808869600j, 
- -0.243692726-  0.200126961j, 
- -0.270986766-  0.460243553j, 
-  0.399368525+  0.249096692j, 
[…]
