### 2017-08-19 13e4d23
Fix issue #36 + pep8 + examples
