+++ src/spectrum/eigenfre.py
@@ -5,7 +5,6 @@ from spectrum import default_NFFT
 from .psd import ParametricSpectrum
 from .tools import twosided_2_onesided, centerdc_2_twosided, twosided_2_centerdc
 
-verbose = False
 
 __all__ = ["music", "ev", "pmusic", "eigen", "pev"]
 
@@ -19,37 +18,56 @@ class pmusic(ParametricSpectrum):
         :width: 80%
         :include-source:
 
-        from spectrum import *
+        from spectrum import pmusic
         p = pmusic(marple_data, 15, NFFT=4096)
         p()
         p.plot()
 
 
+    Another example using a two sinusoidal components::
+
+        n = arange(200)
+        x = cos(0.257*pi*n) + sin(0.2*pi*n) + 0.01*randn(size(n));
+        p = pmusic(x, 6,4)
+        p()
+        p.plot()
+
     """
-    def __init__(self, data, IP, NSIG=None, NFFT=None, sampling=1.):
+    def __init__(self, data, IP, NSIG=None, NFFT=None, sampling=1.,
+        threshold=None, criteria="aic", verbose=False):
         """**Constructor:**
 
         For a detailed description of the parameters, see :func:`arma_estimate`.
 
-        :param array data:     input data (list or numpy.array)
-        :param int IP:
-        :param int NFFT:       total length of the final data sets (padded with zero if needed; default is 4096)
+        :param array data: input data (list or numpy.array)
+        :param int P: maximum number of eigen values to compute. NSIG (if
+            specified) must therefore be less than P.
+        :param int NSIG: If specified, the signal sub space uses NSIG eigen values.
+        :param int NFFT: total length of the final data sets (padded with zero if needed; default is 4096)
         :param float sampling: sampling frequency of the input :attr:`data`.
 
         """
         super(pmusic, self).__init__(data, ar_order=IP,
-                                            NFFT=NFFT, sampling=sampling)
+                                     NFFT=NFFT, sampling=sampling)
         self.NSIG = NSIG
+        self.threshold = threshold
+        self.criteria = criteria
+        self.verbose = verbose
 
     def __call__(self):
-        psd, _eigenvalues = eigen(self.data, self.ar_order, NSIG=self.NSIG,
-            NFFT=self.NFFT, threshold=None, criteria='aic', verbose=verbose, method='music')
+        psd, eigenvalues = eigen(self.data, self.ar_order, NSIG=self.NSIG,
+            NFFT=self.NFFT, threshold=self.threshold, criteria=self.criteria, 
+            verbose=self.verbose, method='music')
+        self.eigenvalues = eigenvalues
+
         if self.datatype == 'real':
             self.psd = twosided_2_onesided(psd)
+            # we need to flip the data
+            self.psd = self.psd[::-1]
         else:
             self.psd = centerdc_2_twosided(psd)
-        self.scale()
 
+        self.scale()
 
     def _str_title(self):
         return "Music PSD estimate\n"
@@ -67,45 +85,53 @@ class pev(ParametricSpectrum):
         :width: 80%
         :include-source:
 
-        from spectrum import *
+        from spectrum import pev
         p = pev(marple_data, 15, NFFT=4096)
         p()
         p.plot()
 
-
     """
-    def __init__(self, data, IP, NSIG=None, NFFT=None, sampling=1.):
+    def __init__(self, data, IP, NSIG=None, NFFT=None, sampling=1.,
+        threshold=None, criteria="aic", verbose=False):
         """**Constructor:**
 
         For a detailed description of the parameters, see :func:`arma_estimate`.
 
[…]
