### 2017-08-19 153340d
Fix examples and pep8
