### 2017-08-19 4caa09b
Fix pep8 and example
