### 2017-08-19 53e2c81
fix examples
