### 2017-08-19 2cf2b70
Fix pep8; remove deprecated hold() function
