+++ doc/create_psd.py
@@ -1,72 +1,74 @@
 import numpy
-from spectrum import *
-from spectrum import ma, arma_estimate
-from spectrum.psd import *
-from spectrum.errors import *
+import spectrum
+from spectrum import tools
+#from spectrum.psd import *
+#from spectrum.errors import *
 from numpy.testing import assert_array_almost_equal
 import pylab
-data = marple_data
+data = spectrum.marple_data
 
 
 def create_all_psd():
+
+
     f = pylab.linspace(0, 1, 4096)
     pylab.clf()
+
+    pylab.figure(figsize=(12,8))
+
     #MA 15 order
-    b, rho = ma(data, 15, 30)
-    psd = arma2psd(B=b, rho=rho)
+    b, rho = spectrum.ma(data, 15, 30)
+    psd = spectrum.arma2psd(B=b, rho=rho)
     newpsd = tools.cshift(psd, len(psd)//2) # switch positive and negative freq
     pylab.plot(f, 10 * pylab.log10(newpsd/max(newpsd)), label='MA 15')
-    pylab.hold(True)
 
     #ARMA 15 order
-    a, b, rho = arma_estimate(data, 15,15, 30)
-    psd = arma2psd(A=a,B=b, rho=rho)
+    a, b, rho = spectrum.arma_estimate(data, 15,15, 30)
+    psd = spectrum.arma2psd(A=a,B=b, rho=rho)
     newpsd = tools.cshift(psd, len(psd)//2) # switch positive and negative freq
     pylab.plot(f, 10 * pylab.log10(newpsd/max(newpsd)), label='ARMA 15,15')
-    #pylab.hold(True)
 
     #yulewalker
-    ar, P,c = aryule(data, 15, norm='biased')
-    psd = arma2psd(A=ar, rho=P)
+    ar, P,c = spectrum.aryule(data, 15, norm='biased')
+    psd = spectrum.arma2psd(A=ar, rho=P)
     newpsd = tools.cshift(psd, len(psd)//2) # switch positive and negative freq
 
     pylab.plot(f, 10 * pylab.log10(newpsd/max(newpsd)), label='YuleWalker 15')
 
     #burg method
-    ar, P,k = arburg(data, order=15)
-    psd = arma2psd(A=ar, rho=P)
+    ar, P,k = spectrum.arburg(data, order=15)
+    psd = spectrum.arma2psd(A=ar, rho=P)
     newpsd = tools.cshift(psd, len(psd)//2) # switch positive and negative freq
-
     pylab.plot(f, 10 * pylab.log10(newpsd/max(newpsd)), label='Burg 15')
 
     #covar method
-    af, pf, ab, pb, pv = arcovar_marple(data, 15)
-    psd = arma2psd(A=af, B=ab, rho=pf)
+    af, pf, ab, pb, pv = spectrum.arcovar_marple(data, 15)
+    psd = spectrum.arma2psd(A=af, B=ab, rho=pf)
     newpsd = tools.cshift(psd, len(psd)//2) # switch positive and negative freq
     pylab.plot(f, 10 * pylab.log10(newpsd/max(newpsd)), label='covar 15')
 
     #modcovar method
-    a, p, pv = modcovar_marple(data, 15)
-    psd = arma2psd(A=a)
+    a, p, pv = spectrum.modcovar_marple(data, 15)
+    psd = spectrum.arma2psd(A=a)
     newpsd = tools.cshift(psd, len(psd)//2) # switch positive and negative freq
     pylab.plot(f, 10 * pylab.log10(newpsd/max(newpsd)), label='modcovar 15')
 
     #correlogram
-    psd = CORRELOGRAMPSD(data, data, lag=15)
[…]
