### 2017-08-19 aa44c68
update doctests
