+++ src/spectrum/linalg.py
@@ -1,15 +1,15 @@
 """
-.. topic:: Linear algebra tools  
+.. topic:: Linear algebra tools
 
     .. autosummary::
-       
+
         csvd
         corrmtx
         pascal
-    
+
     .. codeauthor:: Thomas Cokelaer 2011
- 
-   
+
+
 """
 import numpy
 
@@ -20,16 +20,19 @@ __all__ = ["pascal", "csvd", "corrmtx"]
 
 def pascal(n):
     """Return Pascal matrix
-    
+
     :param int n: size of the matrix
-    
-    >>> pascal(6)
-    array([[   1.,    1.,    1.,    1.,    1.,    1.],
-           [   1.,    2.,    3.,    4.,    5.,    6.],
-           [   1.,    3.,    6.,   10.,   15.,   21.],
-           [   1.,    4.,   10.,   20.,   35.,   56.],
-           [   1.,    5.,   15.,   35.,   70.,  126.],
-           [   1.,    6.,   21.,   56.,  126.,  252.]])
+
+    .. doctest::
+
+        >>> from spectrum import pascal
+        >>> pascal(6)
+        array([[   1.,    1.,    1.,    1.,    1.,    1.],
+               [   1.,    2.,    3.,    4.,    5.,    6.],
+               [   1.,    3.,    6.,   10.,   15.,   21.],
+               [   1.,    4.,   10.,   20.,   35.,   56.],
+               [   1.,    5.,   15.,   35.,   70.,  126.],
+               [   1.,    6.,   21.,   56.,  126.,  252.]])
 
     .. todo:: use the symmetric property to improve computational time if needed
     """
@@ -44,108 +47,108 @@ def pascal(n):
         for i in range(1, n):
             for j in range(1, n):
                 result[i, j] = result[i-1, j] + result[i, j-1]
-    return result      
+    return result
 
 
 
 def corrmtx(x_input, m, method='autocorrelation'):
     r"""Correlation matrix
-    
+
     This function is used by PSD estimator functions. It generates
     the correlation matrix from a correlation data set and a maximum lag.
-    
+
     :param array x: autocorrelation samples (1D)
     :param int m: the maximum lag
-    
-    Depending on the choice of the method, the correlation matrix has different 
-    sizes, but the number of rows is always m+1.  
+
+    Depending on the choice of the method, the correlation matrix has different
+    sizes, but the number of rows is always m+1.
 
     Method can be :
-    
-    * 'autocorrelation': (default) X is the (n+m)-by-(m+1) rectangular Toeplitz 
+
+    * 'autocorrelation': (default) X is the (n+m)-by-(m+1) rectangular Toeplitz
       matrix derived using prewindowed and postwindowed data.
     * 'prewindowed': X is the n-by-(m+1) rectangular Toeplitz matrix derived
       using prewindowed data only.
-    * 'postwindowed': X is the n-by-(m+1) rectangular Toeplitz matrix that 
+    * 'postwindowed': X is the n-by-(m+1) rectangular Toeplitz matrix that
       derived using postwindowed data only.
-    * 'covariance': X is the (n-m)-by-(m+1) rectangular Toeplitz matrix       
+    * 'covariance': X is the (n-m)-by-(m+1) rectangular Toeplitz matrix
       derived using nonwindowed data.
     * 'modified': X is the 2(n-m)-by-(m+1) modified rectangular Toeplitz
-      matrix that generates an autocorrelation estimate for the length n data 
+      matrix that generates an autocorrelation estimate for the length n data
[…]
