+++ src/spectrum/linear_prediction.py
@@ -7,7 +7,7 @@ from .levinson import LEVINSON, rlevinson
 import numpy
 from scipy.signal import deconvolve
 
-__all__ = ['ac2poly', 'poly2ac', 'ac2rc', 'rc2poly', 'rc2ac', 'is2rc', 'rc2is', 
+__all__ = ['ac2poly', 'poly2ac', 'ac2rc', 'rc2poly', 'rc2ac', 'is2rc', 'rc2is',
     'rc2lar', 'lar2rc', 'poly2rc', 'lsf2poly', 'poly2lsf']
 
 """
@@ -20,7 +20,7 @@ def ac2poly(data):
     """Convert autocorrelation sequence to prediction polynomial
 
     :param array data:    input data (list or numpy.array)
-    :return: 
+    :return:
         * AR parameters
         * noise variance
 
@@ -32,16 +32,18 @@ def ac2poly(data):
 
     .. doctest::
 
-        >>> r = [5, -2, 1]
+        >>> from spectrum import ac2poly
+        >>> from numpy import array
+        >>> r = [5, -2, 1.01]
         >>> ar, e = ac2poly(r)
         >>> ar
-        array([ 1. ,  0. , -0.2])
+        array([ 1.  ,  0.38, -0.05])
         >>> e
-        4.7999999999999998
+        4.1895000000
 
     """
     a, e, _c = LEVINSON(data)
-    a = numpy.insert(a, 0, 1) 
+    a = numpy.insert(a, 0, 1)
     return a, e
 
 
@@ -59,39 +61,44 @@ def ac2rc(data):
     """
     a, e, _c = LEVINSON(data)
     return a, data[0]
-    
-    
+
+
 def poly2ac(poly, efinal):
     """ Convert prediction filter polynomial to autocorrelation sequence
 
     :param array poly: the AR parameters
     :param efinal: an estimate of the final error
-    :return: the autocorrelation  sequence
+    :return: the autocorrelation  sequence in complex format.
 
     .. doctest::
 
-        >>> ar = [ 1. ,  0. , -0.2]
-        >>> efinal = 4.8
-        >>> r = poly2ac(ar, efinal)
+        >>> from numpy import array
+        >>> from spectrum import poly2ac
+        >>> poly = [ 1. ,  0.38 , -0.05]
+        >>> efinal = 4.1895
+        >>> poly2ac(poly, efinal)
+        array([ 5.00+0.j, -2.00+0.j,  1.01-0.j])
 
     """
     results = rlevinson(poly, efinal)
-    return results[0] 
-    
+    return results[0]
+
+
 def ar2rc(ar):
     """ Convert autoregressive parameters into reflection coefficients """
     raise NotImplementedError
-    
-    
+
+
 def poly2rc(a, efinal):
     """Convert prediction filter polynomial to reflection coefficients
 
     :param a: AR parameters
-    :param efinal: 
+    :param efinal:
     """
     results = rlevinson(a, efinal)
     return results[2]
-    
+
+
 def rc2poly(kr, r0=None):
     """convert reflection coefficients to prediction filter polynomial
 
@@ -114,7 +121,7 @@ def rc2poly(kr, r0=None):
 
     e[0] = e0 * (1. - numpy.conj(numpy.conjugate(kr[0])*kr[0]))
 
-    # Continue the recursion for k=2,3,...,p, where p is the order of the 
+    # Continue the recursion for k=2,3,...,p, where p is the order of the
     # prediction polynomial.
 
     for k in range(1, p):
@@ -144,17 +151,17 @@ def is2rc(inv_sin):
 
[…]
