+++ test/test_transfer.py
@@ -15,7 +15,6 @@ def test_tf2zpk():
     assert_almost_equal(p[2], 0.15838444+0.j)
 
 
-
 def test_tf2zp():
     b = [2, 3, 0]
     a = [1, 0.4, 1]
@@ -24,5 +23,75 @@ def test_tf2zp():
     assert k == 2
     assert_almost_equal(p[0], -0.2000 + 0.9798j,4)
     assert_almost_equal(p[1],  -0.2000 - 0.9798j,4)
+
+    transfer.zp2tf(z,p,k)
+
+def test_eqtlength():
+    a, b = transfer.eqtflength([1,2,3,4], np.array([1,2]))
+    assert all(b == np.array([1,2,0,0]))
+    a, b = transfer.eqtflength([1,2,3,4], [1,2])
+    assert b == [1,2,0,0]
+    a, b = transfer.eqtflength(np.array([1,2]), [1,2,3,4])
+    assert all(a == np.array([1,2,0,0]))
+    a, b = transfer.eqtflength([1,2], [1,2,3,4])
+    assert a == [1,2,0,0]
+    a, b = transfer.eqtflength([1,2], [1,2])
+    assert a == b
+
+
+def test_latc2tf():
+    try:
+        transfer.latc2tf()
+        assert False
+    except NotImplementedError:
+        assert True
+    except:
+        assert False
+
+
+def test_latcfilt():
+    try:
+        transfer.latcfilt()
+        assert False
+    except NotImplementedError:
+        assert True
+    except:
+        assert False
+
+
+def test_tf2ss():
+    try:
+        transfer.tf2ss()
+        assert False
+    except NotImplementedError:
+        assert True
+    except:
+        assert False
+
+
+def test_tf2sos():
+    try:
+        transfer.tf2sos()
+        assert False
+    except NotImplementedError:
+        assert True
+    except:
+        assert False
+
+
+def test_ss2zpk():
+    z,p,k = [1.5,1], [-0.2,1], 2.
+
+    zp,pp,kp = transfer.ss2zpk(*transfer.zpk2ss(z,p,k))
+    #assert zp == z
+    # FIXME: this fails
+    #assert pp == p
+    assert k == kp
+
+
+
+
+
+
 
 
