### 2017-08-20 737faf4
remove useless _correlogram method
