+++ doc/source/contributions.rst
@@ -9,9 +9,8 @@ Pleas see https://github.com/cokelaer/spectrum/graphs/contributors for details.
 How to contribute ?
 -----------------------
 
-If you have bug, issues, contribute, please join the development on https://github.com/cokelaer/spectrum.
-
-
+If you have found a bug, have issues or contributions, please join the 
+development on https://github.com/cokelaer/spectrum.
 
 
 Some notebooks
