+++ doc/source/ChangeLog.rst
@@ -6,15 +6,17 @@ Version 0.6
 
 * 0.6.7:
 
-   * refactored the requirements files (add a requirements-dev.txt) and 
-     update the documentation (installation)      accordingly
-   * BUG fixes: 
+    * refactored the requirements files (add a requirements-dev.txt) and 
+      update the documentation (installation)      accordingly
+    * BUG fixes: 
        * correlogram: real-data case had the data flipped
        * pmusic/pev: real-data case had the data flipped
        * fix the AKICc criteria code
-   * Updates:
+    * Updates:
        * pmusic/pev: add the threshold and criteria arguments
        * more tests for the criteria and eigenfre modules
+    * Changes:
+       * Spectrum class: remove _correlogram method (use pcorrelogram instead)
 
 * 0.6.6:
 
