+++ doc/source/index.rst
@@ -1,7 +1,38 @@
 
 .. topic:: Spectrum: a Spectral Analysis Library in Python
 
-    .. include:: overview.txt
+.. image:: https://badge.fury.io/py/spectrum.svg
+    :target: https://pypi.python.org/pypi/spectrum
+
+.. image:: https://secure.travis-ci.org/cokelaer/spectrum.png
+    :target: http://travis-ci.org/cokelaer/spectrum
+
+.. image:: https://coveralls.io/repos/cokelaer/spectrum/badge.png?branch=master 
+    :target: https://coveralls.io/r/cokelaer/spectrum?branch=master 
+
+.. image:: https://landscape.io/github/cokelaer/spectrum/master/landscape.png
+    :target: https://landscape.io/github/cokelaer/spectrum/master
+
+.. image:: https://anaconda.org/conda-forge/spectrum/badges/license.svg
+      :target: https://anaconda.org/conda-forge/spectrum
+
+.. image:: https://anaconda.org/conda-forge/spectrum/badges/installer/conda.svg
+      :target: https://conda.anaconda.org/conda-forge
+
+.. image:: https://anaconda.org/conda-forge/spectrum/badges/downloads.svg
+      :target: https://anaconda.org/conda-forge/spectrum
+
+
+:contributions: Please join https://github.com/cokelaer/spectrum
+:contributors: https://github.com/cokelaer/spectrum/graphs/contributors
+:issues: Please use https://github.com/cokelaer/spectrum/issues
+:documentation: http://pyspectrum.readthedocs.io/ 
+
+
+
+.. include:: overview.txt
+
+
 
 
 
