### 2017-08-20 d0cbb3b
add back run() for back compatibility
