### 2017-08-20 e1d94cd
fix doctest
