### 2017-08-20 02e26f2
remove the p() and p.run() calls
