+++ src/spectrum/psd.py
@@ -1,4 +1,6 @@
 """This module provides the Base class for PSDs"""
+import logging
+
 import numpy
 
 from spectrum.tools import nextpow2
@@ -8,7 +10,6 @@ from spectrum import tools
 
 
 __all__ = ["Spectrum", "FourierSpectrum", "ParametricSpectrum"]
-debug = True
 
 
 class Range(object):
@@ -102,7 +103,7 @@ class Range(object):
 
         ::
 
-            >>> print list(Range(8).centerdc_gen())
+            >>> print(list(Range(8).centerdc_gen()))
             [-0.5, -0.375, -0.25, -0.125, 0.0, 0.125, 0.25, 0.375]
 
         """
@@ -114,7 +115,7 @@ class Range(object):
 
         ::
 
-            >>> print list(Range(8).centerdc_gen())
+            >>> print(list(Range(8).centerdc_gen()))
             [0, 0.125, 0.25, 0.375, 0.5, 0.625, 0.75, 0.875]
 
         """
@@ -129,9 +130,9 @@ class Range(object):
 
         ::
 
-            >>> print list(Range(8).onesided())
+            >>> print(list(Range(8).onesided()))
             [0.0, 0.125, 0.25, 0.375, 0.5]
-            >>> print list(Range(9).onesided())
+            >>> print(list(Range(9).onesided()))
             [0.0, 0.1111, 0.2222, 0.3333, 0.4444]
 
         """
@@ -214,7 +215,7 @@ class Spectrum(object):
     At any time, you can get general information about the Spectrum instance::
 
         >>> p = Spectrum(marple_data)
-        >>> print p
+        >>> print(p)
         Spectrum summary
             Data length is 64
             PSD not yet computed
@@ -309,7 +310,7 @@ class Spectrum(object):
         # e.g. used in this external notebook:
         # http://nbviewer.jupyter.org/gist/juhasch/5182528
         self()
-    
+
     def _getDetrend(self):
         return self.__detrend
     def _setDetrend(self, detrend):
@@ -346,18 +347,18 @@ class Spectrum(object):
         return self.__NFFT
     def _setNFFT(self, NFFT):#if NFFT is changed, we need to redo the padding
         if self.__NFFT == NFFT and self.__NFFT is not None:
-            #print 'NFFT is the same, nothing to do'
+            logging.debug('NFFT is the same, nothing to do')
             return
         new_nfft = None
         if NFFT == 'nextpow2':
             n = nextpow2(self.data.size)
             new_nfft = int(pow(2, n))
-            #print('NFFT is based on nextpow2: %s' % new_nfft)
+            logging.debug('NFFT is based on nextpow2: %s' % new_nfft)
         elif NFFT is None:
-            #print 'NFFT set to data length',
+            logging.debug('NFFT set to data length')
             new_nfft = self.N
         elif isinstance(NFFT, int):
-            #print 'NNFT set  manually to',
+            logging.debug('NNFT set  manually to')
             assert NFFT > 0, 'NFFT must be a positive integer'
             new_nfft = NFFT
         else:
@@ -410,7 +411,7 @@ class Spectrum(object):
             newpsd = self.get_converted_psd(sides)
             self.__psd = newpsd
         self.__sides = sides
-        #print '------------> %s %s' % (self.__sides, sides)
+        logging.debug('------------> %s %s' % (self.__sides, sides))
         # we set the PSD by hand, so we can consider that PSD is up-to-date
         self.modified = False
     _doc_sides = """Getter/Setter to the :attr:`sides` attributes.
@@ -451,8 +452,7 @@ class Spectrum(object):
 
     def _getPSD(self):
         if self.__psd is None:
[…]
