+++ src/spectrum/modcovar.py
@@ -8,7 +8,7 @@ __all__ = ['modcovar', 'modcovar_marple', 'pmodcovar']
 def modcovar_marple (X,IP):
     """Fast algorithm for the solution of the modified covariance least squares normal equations.
 
-    This implementation is based on [Marple]_. This code is far more 
+    This implementation is based on [Marple]_. This code is far more
     complicated and slower than :func:`modcovar` function, which is now the official version.
     See :func:`modcovar` for a detailed description of Modified Covariance method.
 
@@ -25,18 +25,18 @@ def modcovar_marple (X,IP):
               3. if P' is not a positive value
               4. if DELTA and GAMMA do not lie in the range 0 to 1
 
-    
-    :validation: the AR parameters are the same as those returned by 
+
+    :validation: the AR parameters are the same as those returned by
         a completely different function :func:`modcovar`.
-        
-    .. note:: validation. results similar to test example in Marple but 
-        starts to differ for    ip~8. with ratio of 0.975 for ip=15 probably 
+
+    .. note:: validation. results similar to test example in Marple but
+        starts to differ for    ip~8. with ratio of 0.975 for ip=15 probably
         due to precision.
 
-        
+
     :References: [Marple]_
     """
-    Pv = [] 
+    Pv = []
     N = len(X)
     A = numpy.zeros(N, dtype=complex)
     D = numpy.zeros(N, dtype=complex)
@@ -55,7 +55,7 @@ def modcovar_marple (X,IP):
     LAMBDA = (X[0] * X[N-1]).conjugate()*R4
     C[0] = X[N-1] * R4
     D[0] = X[0].conjugate() * R4
-      
+
     M = 0
     if (IP ==0):
         P = (.5*R1+R2+R3)/float(N)
@@ -85,8 +85,8 @@ def modcovar_marple (X,IP):
                 #print 'withini loop', K, THETA, PSI, XI, R[K], SAVE1
                 #print 'x   -------', C[K], C[K+1]
                 #print 'x   -------', D[K], D[K+1]
-                #print 'x   -------', N-K-2, K+1,N-M-1,  M, M-K-1 
-                #print 'x   -------', X[N-K-2], X[K+1],X[N-M-1],  X[M], X[M-K-1] 
+                #print 'x   -------', N-K-2, K+1,N-M-1,  M, M-K-1
+                #print 'x   -------', X[N-K-2], X[K+1],X[N-M-1],  X[M], X[M-K-1]
 
 
         #  Order update of A vector
@@ -200,7 +200,7 @@ def modcovar_marple (X,IP):
         #print R3, R4, P, DELTA, GAMMA, LAMBDA
         #print 'Afinal=',A[0:2]
         #print 'P=',P
-        if (P > 0.): 
+        if (P > 0.):
             pass
         else:
             raise ValueError("Found a invalid negative P value ")
@@ -211,12 +211,12 @@ def modcovar_marple (X,IP):
 
 
 
- 
+
 
 
 def modcovar(x, order):
     """Simple and fast implementation of the covariance AR estimate
-    
+
     This code is 10 times faster than :func:`modcovar_marple` and more importantly
     only 10 lines of code, compared to a 200 loc for :func:`modcovar_marple`
 
@@ -226,12 +226,12 @@ def modcovar(x, order):
     :return:
         * P    - Real linear prediction variance at order IP
         * A    - Array of complex linear prediction coefficients
-    
+
 
     .. plot::
         :include-source:
         :width: 80%
-        
+
         from spectrum import *
         from pylab import *
 
@@ -240,72 +240,71 @@ def modcovar(x, order):
[…]
