+++ src/spectrum/burg.py
@@ -90,7 +90,6 @@ class pburg(ParametricSpectrum):
 
         from spectrum import *
         p = pburg(marple_data, 15, NFFT=4096)
-        p()
         p.plot(sides='centerdc')
 
 
