### 2017-08-20 a4ac130
update doc and tests
