### 2017-08-20 cokelaer/spectrum#35 @cokelaer
done on the master branch. Documentation updated
