### 2017-08-20 cokelaer/spectrum#33 @cokelaer
I believe this is now done in moist of the documentation and examples
