+++ AUTHORS.rst
@@ -1,0 +1,7 @@
+- Thomas Cokelaer (main author)
+- https://github.com/juhasch
+- https://github.com/anielsen001
+- https://github.com/carlkl
+- https://github.com/gozzilli
+- https://github.com/wanglongqi
+
