### 2017-08-20 18ad815
Fix the Manisfest
