### 2017-08-26 74fad8e
divided PSD by T (instead of multipliying).This should fix issue #38
