### 2017-08-26 d184edf
. remove call to scale() and do not change PSD at f=0; This should fix #38. Same results as octave/matlab on a real-case example
