### 2017-08-26 f068b52
some logging and cleaning
