+++ doc/source/autosum.rst
@@ -12,6 +12,7 @@ Reference guide
     ref_others.rst
 
 
+
 .. toctree::
     :hidden:
 
