+++ doc/source/ref_others.rst
@@ -8,18 +8,17 @@ Tools and classes
 
 Classes
 ---------
+.. module:: psd
 .. automodule:: spectrum.psd
     :members:
     :undoc-members:
     :synopsis: psd classes to ease test and manipulation of PSDs
 
 
-
-
 Correlation
 -------------
 
-#.. module:: correlation
+.. module:: correlation
 .. automodule:: spectrum.correlation
     :members:
     :undoc-members:
@@ -28,8 +27,8 @@ Correlation
 Tools
 ----------
 
-#.. module:: tools
-#    :synopsis: utilities to manipulate PSDs
+.. module:: tools
+    :synopsis: utilities to manipulate PSDs
 .. automodule:: spectrum.tools
     :members:
     :undoc-members:
@@ -38,7 +37,7 @@ Tools
 datasets
 -----------
 
-#.. module:: datasets
+.. module:: datasets
 .. automodule:: spectrum.datasets
     :members:
     :undoc-members:
@@ -65,8 +64,8 @@ eigen
 levinson
 ~~~~~~~~
 
-#.. module:: levinson
-#    :synopsis: levinson recursion tools
+.. module:: levinson
+    :synopsis: levinson recursion tools
 .. automodule:: spectrum.levinson
     :members:
     :undoc-members:
@@ -83,8 +82,8 @@ toeplitz
 linalg
 ~~~~~~~~~
 
-#.. module:: linalg
-#    :synopsis: Linear algebra tools
+.. module:: linalg
+    :synopsis: Linear algebra tools
 .. automodule:: spectrum.linalg
     :members:
     :undoc-members:
