+++ src/spectrum/levinson.py
@@ -10,10 +10,13 @@
 import numpy
 
 
+__all__ = ["LEVINSON", "rlevinson"]
+
+
 def LEVINSON(r, order=None, allow_singularity=False):
     r"""Levinson-Durbin recursion.
 
-    Find the coefficients of a length(r)-1 order autoregressive linear process 
+    Find the coefficients of a length(r)-1 order autoregressive linear process
 
     :param r: autocorrelation sequence of length N + 1 (first element being the zero-lag autocorrelation)
     :param order: requested order of the autoregressive coefficients. default is N.
@@ -22,33 +25,33 @@ def LEVINSON(r, order=None, allow_singularity=False):
     :return:
         * the `N+1` autoregressive coefficients :math:`A=(1, a_1...a_N)`
         * the prediction errors
-        * the `N` reflections coefficients values 
-    
-    This algorithm solves the set of complex linear simultaneous equations 
-    using Levinson algorithm.  
+        * the `N` reflections coefficients values
+
+    This algorithm solves the set of complex linear simultaneous equations
+    using Levinson algorithm.
 
     .. math::
 
-        \bold{T}_M \left( \begin{array}{c} 1 \\ \bold{a}_M \end{array} \right) = 
+        \bold{T}_M \left( \begin{array}{c} 1 \\ \bold{a}_M \end{array} \right) =
         \left( \begin{array}{c} \rho_M \\ \bold{0}_M  \end{array} \right)
 
     where :math:`\bold{T}_M` is a Hermitian Toeplitz matrix with elements
-    :math:`T_0, T_1, \dots ,T_M`. 
-    
-    .. note:: Solving this equations by Gaussian elimination would 
-        require :math:`M^3` operations whereas the levinson algorithm 
+    :math:`T_0, T_1, \dots ,T_M`.
+
+    .. note:: Solving this equations by Gaussian elimination would
+        require :math:`M^3` operations whereas the levinson algorithm
         requires :math:`M^2+M` additions and :math:`M^2+M` multiplications.
-  
-    This is equivalent to solve the following symmetric Toeplitz system of 
+
+    This is equivalent to solve the following symmetric Toeplitz system of
     linear equations
-    
+
     .. math::
-    
+
         \left( \begin{array}{cccc}
         r_1 & r_2^* & \dots & r_{n}^*\\
         r_2 & r_1^* & \dots & r_{n-1}^*\\
         \dots & \dots & \dots & \dots\\
-        r_n & \dots & r_2 & r_1 \end{array} \right)    
+        r_n & \dots & r_2 & r_1 \end{array} \right)
         \left( \begin{array}{cccc}
         a_2\\
         a_3 \\
@@ -60,11 +63,11 @@ def LEVINSON(r, order=None, allow_singularity=False):
         -r_3 \\
         \dots \\
         -r_{N+1}  \end{array} \right)
-         
-    where :math:`r = (r_1  ... r_{N+1})` is the input autocorrelation vector, and 
-    :math:`r_i^*` denotes the complex conjugate of :math:`r_i`. The input r is typically 
-    a vector of autocorrelation coefficients where lag 0 is the first 
-    element :math:`r_1`. 
+
+    where :math:`r = (r_1  ... r_{N+1})` is the input autocorrelation vector, and
+    :math:`r_i^*` denotes the complex conjugate of :math:`r_i`. The input r is typically
+    a vector of autocorrelation coefficients where lag 0 is the first
+    element :math:`r_1`.
 
 
     .. doctest::
@@ -78,23 +81,23 @@ def LEVINSON(r, order=None, allow_singularity=False):
     T0  = numpy.real(r[0])
     T = r[1:]
     M = len(T)
-    
+
     if order is None:
-        M = len(T) 
+        M = len(T)
     else:
[…]
