### 2017-08-26 02a3ed3
add object in class parent
