### 2017-08-26 8e76fb9
add scale_by_freq argument in the constructor
