### 2017-08-26 002216d
cleaning and mre tests
