+++ src/spectrum/covar.py
@@ -1,10 +1,13 @@
+import logging
+
 import numpy
 from .psd import ParametricSpectrum
 
 
 __all__ = ["arcovar", "arcovar_marple", "pcovar", 'arcovar_marple']
 
-def arcovar_marple(x, order, debug=False):
+
+def arcovar_marple(x, order):
     r"""Estimate AR model parameters using covariance method
 
     This implementation is based on [Marple]_. This code is far more
@@ -75,29 +78,26 @@ def arcovar_marple(x, order, debug=False):
     #ip +1 because we want to enter in the loop to run the first part of the code.
     pbv = []
     for m in range(0, order+1):
-        if debug:
-            print(('----------------------------m=', m))
-            print((c[0:2]))
-            print((d[0:2]))
+        logging.debug('----------------------------m=', m)
+        logging.debug(c[0:2])
+        logging.debug(d[0:2])
         r1 = 1./pf
         r2 = 1./pb
         r3 = 1./delta
         r4 = 1./gamma
-        if debug:
-            print(('starting r1r2r3r4=', r1, r2, r3, r4, pf, pb, delta, gamma))
+
+        #logging.debug('starting r1r2r3r4=', r1, r2, r3, r4, pf, pb, delta, gamma)
+
         #Order update: AF and AB vectors ; time update: C and D vectors
         temp = 0.+0.j
         for k in range(m+1, N):
             temp = temp + x[k]*x[k-m-1].conjugate()
 
-        if debug:
-            print(('temp=', temp))
         r[m] =  temp.conjugate()
         theta = x[0] * c[m]
-        if debug:
-            print(('theta', theta))
-            print(('cccccccccc', c[0:2]))
-            print(('dddddd', d[0:2]))
+        #print(('theta', theta))
+        #    print(('c=', c[0:2]))
+        #    print(('d=', d[0:2]))
         if m == 0:
             pass
         else:
@@ -108,31 +108,32 @@ def arcovar_marple(x, order, debug=False):
                 #print 'loop1 k=', k
                 #print '         theta=',theta, 'r[k]=',r[k], 'temp=', temp
                 #print '         c=',c[k], 'af=',af[m-k-1]
-        if m>0:
+        """if m > 0:
             if debug:
                 print((m, N-m))
                 print(('Xk=0',x[m-0],x[N-m-1], x[N-m+0]))
-            if m>1:
+            if m > 1:
                 if debug:
                     print('Xk=1',x[m-1],x[N-m-1], x[N-m+1])
+        """
         c1 = -temp * r2
         c2 = -r1 * temp.conjugate()
         c3 = theta * r3
         c4 = r4 *theta.conjugate()
-        if debug:
-            print('c1c2c3c4 before af=',c1 ,c2 ,c3 ,c4)
+        #if debug:
+        #    print('c1 c2 c3 c4 before af=',c1 ,c2 ,c3 ,c4)
         af[m] = c1                    #             ! Eq. (8.C.19)
         ab[m] = c2                    #             ! Eq. (8.C.22)
         save = c[m]
         c[m] = save + c3*d[m]
         d[m] = d[m] + c4*save
-        if debug:
-            print('res',m,'af[m]=',af[m], ab[m], save, 'temp=',temp)
+        #if debug:
+        #    print('res',m,'af[m]=',af[m], ab[m], save, 'temp=',temp)
 
         if m == 0:
             pass
         else:
-            if debug:print('af before', af[0:2])
+            #if debug:print('af before', af[0:2])
             for k in range(0, m):
                 save = af[k]
                 af[k] = save + c1 * ab[m-k-1] # Eq. (8.C.18)
@@ -141,16 +142,15 @@ def arcovar_marple(x, order, debug=False):
[…]
