+++ test/test_spectrum.py
@@ -4,3 +4,4 @@ from spectrum import *
 
 def test_spectrum():
     assert spectrum.default_NFFT
+    spectrum_set_level("DEBUG")
