### 2017-08-26 c43f07b
increase version 0.7.0
