### 2017-08-26 50dc40d
Fix all parametric method (real case) by adding the scale_by_freq argument and fixing the f=0 and f=Fs/2 cases
