+++ examples/README.txt
@@ -1,0 +1,4 @@
+Gallery
+==========
+
+General-purpose examples for spectrum library. 
