### 2017-08-26 91a2961
add one more example in the gallery and update doc
