+++ doc/source/conf.py
@@ -171,7 +171,13 @@ def setup(app):
 
 # The theme to use for HTML and HTML Help pages.  Major themes that come with
 # Sphinx are currently 'default' and 'sphinxdoc'.
-html_theme = "standard"
+on_rtd = os.environ.get("READTHEDOCS", None) == "True"
+if not on_rtd:
+    import sphinx_rtd_theme
+    html_theme = 'sphinx_rtd_theme'
+    html_theme_path = [sphinx_rtd_theme.get_html_theme_path()]
+else:
+    html_theme = "default"
 
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
@@ -203,6 +209,9 @@ html_logo = os.path.join('logo.png')
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
+html_static_path = ['_static']
+
+
 #html_static_path = ['_static', os.path.join(extensionPath, '_static')]
 #html_style = 'mine.css'
 # If not '', a 'Last updated on:' timestamp is inserted at every page bottom,
