### 2017-08-27 612ac0a
add sphinx_gallery for RTD
