### 2017-08-27 489a511
add sphinx_gallery for RTD
