+++ doc/source/gallery.rst
@@ -1,0 +1,22 @@
+
+.. raw:: html
+
+   <div class="body">
+   <div id="index-grid" class="section group">
+   <div class="col span_1_of_3">
+        <h3><a href="install.html">Installation</a></h3>
+        <p>conda install spectrum</p>
+        <h3><a href="auto_examples/index.html">Examples</a></h3>
+        <p>Visit our example gallery or jump to the main documentation</p>
+    </div>
+
+
+.. raw:: html
+
+            </div>
+        </p>
+        </div>
+        </div>
+   </div>
+   <div style="clear: left"></div>
+
