### 2017-08-27 adb17ec
fix pma real case (Fs=0 and Fs/2)
