### 2017-08-27 55b7846
Fix Fs=0 and Fs/2 real case for pcovar
