### 2017-08-27 5936ed0
cleanup (remove plot()
