### 2017-08-27 d0dce7d
update tests
