### 2017-08-27 e32cec8
update doc/ pep8 / use logging in eigenfre
