### 2017-08-27 0e169d6
Fix link in the documentation
