### 2017-08-27 f127044
Update paper.md
