### 2017-08-27 e84c4d2
Update paper.md
