### 2017-08-27 9b54427
set cmap in plotting only
