### 2017-08-27 8d8ed5e
add spectrum_data function
