+++ src/spectrum/__init__.py
@@ -6,11 +6,9 @@ def spectrum_set_level(level):
     logging.getLogger().setLevel(level)
 
 
-#: default number os samples used to compute FFT
+#: default number of samples used to compute FFT
 default_NFFT = 4096
 
-from .mtm import *
-from .transfer import *
 from .arma import *
 from .burg import *
 from .cholesky import *
@@ -18,8 +16,10 @@ from .correlation import *
 from .correlog import *
 from .covar import *
 from .criteria import *
+from .datasets import *
 from .eigen import *
 from .eigenfre import *
+from .io import *
 from .levinson import *
 from .linear_prediction import *
 from .linalg import *
@@ -27,13 +27,14 @@ from .linalg import *
 from .lpc import *
 from .minvar import *
 from .modcovar import *
+from .mtm import *
 from .periodogram import *
 from .psd import *
-from .datasets import *
+from .spectrogram import *
 from .tools import *
 from .toeplitz import *
+from .transfer import *
 from .window import *
 from .waveform import *
 from .yulewalker import *
-
 
