### 2017-08-27 c6358b1
fix modifier in the psd computation. prevent NFFT from being changed; fix get_converted_psd; use logging
