### 2017-08-27 bedf287
add MultiTapering class
