### 2017-08-27 6d47480
set scale_by_freq to False by default for consistency)
