### 2017-08-27 5ef2c97
Fix the 2D case for speriodogram
