### 2017-08-27 858667f
update all tests reaching 98% coverage !
