+++ test/test_psd.py
@@ -6,7 +6,7 @@ from spectrum.errors import *
 from numpy.testing import assert_array_almost_equal
 import pylab
 data = marple_data
-
+from easydev import TempFile
 
 def test_psd_module_range():
     N = 10
@@ -30,7 +30,6 @@ def test_psd_module_range():
     r.onesided()
     r.twosided()
     print(r)
-
 
 
 def test_spectrum_simple():
@@ -74,12 +73,12 @@ def test_spectrum_simple():
     s.modified = False
     s.data = [1,2,3,4]
     assert s.modified is True
-   
+
     # test the data_y attribute
     assert s.data_y is None
     s.data_y = [1,2,3,4]
     assert s.data_y == [1,2,3,4]
-    
+
     try:
         s.frequencies("dummy") 
         assert False
@@ -104,7 +103,7 @@ def test_spectrum_simple():
     except:
         assert True
 
-    p = parma(data_two_freqs(), 8,4,15)
+    p = parma(data_two_freqs(), 8, 4, 15)
     try:
         p.plot()
         assert False
@@ -112,60 +111,73 @@ def test_spectrum_simple():
         assert True
 
     # power method
-    p = parma(marple_data, 8,4,10)
+    p = parma(marple_data, 8, 4, 10)
     p()
     assert p.power() < 7381 and p.power() > 7380
     p.scale_by_freq = True
-    assert p.power() > 0.286 and p.power() < 0.287
+    assert p.power() > 115.3253 and p.power() < 115.3254
 
 
+def test_psd_modified():
+    data = [1,2,3,4,6,7,8,9,10,1,2,3,4,5,6,7,8]
+    s = parma(data, 4,2,4)
+    s() 
+    assert s.modified is False
+    s.NFFT = s.NFFT
+    assert s.modified is False
 
-class test_spectrum():
-    def __init__(self):
-        #create a spectrum
-        self.data = [1,2,3,4]
-        self.init()
-        assert self.s.N == 4
+    s.NFFT = 4096
+    assert s.modified is True
+    s.psd
+    assert s.modified is False
 
-    def init(self):
-        self.s = Spectrum(self.data)
-        print(self.s) #__str__ when psd is not yet computed
+    s.scale_by_freq = False
+    assert s.modified is False
 
-    def test_sides(self):
-        """Functional tests on p.sides = ..."""
-        # test all sides cases by starting from a side, converting to anothre
-        # and convert back to the original
-        self.init()
-        assert self.s.NFFT == 4
-        assert self.s.sides == 'onesided'
-        self.s.psd = [10, 2, 3, 4, 6]
-        for x in ['onesided', 'twosided', 'centerdc']:
-            for y in ['onesided', 'twosided', 'centerdc']:
-                if x != y:
-                    self.s.sides = x
-                    p0 = self.s.psd.copy()
-                    self.s.sides = y
-                    self.s.sides = x
-                    assert_array_almost_equal(self.s.psd, p0)
+    s.get_converted_psd("onesided")
+    try:
+        s.get_converted_psd("dummy")
+        assert False
+    except:
+        assert True
 
-    def test_frequencies(self):
-        """Functional tests one s.frequencies(...)"""
-        self.init()
-        self.s.frequencies('onesided')
-        self.s.frequencies('twosided')
-        self.s.frequencies('centerdc')
-        try:
-            self.s.frequencies("dummy")
-            assert False
-        except:
-            assert True
 
-    def test_set_psd(self):
-        self.init()
-        self.s.psd = [1,2,3]
-        assert self.s.NFFT == 4
[…]
