+++ test/test_mtm.py
@@ -15,8 +15,28 @@ def test_pmtm():
 
 
     res = pmtm(data, 2.5, show=False, method="eigen")
+    res = pmtm(data, 2.5, show=False, method="unity")
     res = pmtm(data, 2.5, method="eigen", show=True)
+    res = pmtm(data, 2.5, method="adapt", show=True)
     #res = pmtm(data, 2.5, show=False, method="eigen", show=True)
+
+    # e and v must be provided together
+    try:
+        res = pmtm(data, 2.5, show=False, e=1, v=None)
+        assert False
+    except:
+        assert True
+
+    # provide v and e
+    v,e = dpss(64,4,2)
+    pmtm(marple_data, NW=4, k=2, v=v, e=e);
+
+
+    try:
+        pmtm(marple_data, NW=None, k=2);
+        assert False
+    except:
+        assert True
 
 
 def test_fftconvolve():
@@ -36,3 +56,12 @@ def test_crosscorr():
 
 def test_mtm():
     mtm._other_dpss_method(64,4,10)
+
+
+def test_Multitapering():
+    p = MultiTapering(data_two_freqs(), 4,2)
+    p()
+    p = MultiTapering(marple_data, 4,2, method="eigen")
+    p()
+    print(p)
+    p._str_title()
