### 2017-08-27 d358508
fix develop/install mode of the spectrum_data function
