### 2017-08-27 5961aab
fix dev/install mode
