+++ doc/source/overview.txt
@@ -8,11 +8,16 @@
 
 **Spectrum** is a Python library that contains tools to estimate Power Spectral Densities based on Fourier transform, Parametric methods or eigenvalues analysis:
 
-    * The Fourier methods are based upon correlogram, periodogram and Welch estimates. Standard tapering windows (Hann, Hamming, Blackman) and more exotic ones are available (DPSS, Taylor, ...). 
-    * The parametric methods are based on Yule-Walker, BURG, MA and ARMA, covariance and modified covariance methods.
-    * Non-parametric methods based on eigen analysis (e.g., MUSIC) and minimum variance analysis are also implemented.
-    * Multitapering is also available
+* The Fourier methods are based upon correlogram, periodogram and 
+  Welch estimates. Standard tapering windows (Hann, Hamming, Blackman) 
+  and more exotic ones are available (DPSS, Taylor, ...). 
+* The parametric methods are based on Yule-Walker, BURG, MA and ARMA, 
+  covariance and modified covariance methods.
+* Non-parametric methods based on eigen analysis (e.g., MUSIC) and minimum 
+  variance analysis are also implemented.
+* Finally, Multitapering combines several orthogonal tapering windows.
 
+--
 
 The targetted audience is diverse. Although the use of power spectrum of a
 signal is fundamental in electrical engineering (e.g. radio communications,
