+++ setup.py
@@ -14,11 +14,6 @@ version              = '%d.%d.%d' % (_MAJOR, _MINOR, _MICRO)
 release              = '%d.%d' % (_MAJOR, _MINOR)
 
 
-data_files = [ (
-        pj('share', 'data'), 
-        glob.glob(os.path.join('share','data', '*'))
-    )]
-
 with open('README.rst') as f:
     readme = f.read()
 
