### 2017-08-29 eadfc36
Update paper.md
