+++ doc/joss/paper.bib
@@ -38,4 +38,37 @@
   title = {The use of fast Fourier transform for the estimation of power spectra: a method based on time averaging over short, modified periodograms}
   year = 1967
   }
- 
+
+@Misc{scipy:2001,
+  author =    {Eric Jones and Travis Oliphant and Pearu Peterson and others},
+  title =     {{SciPy}: Open source scientific tools for {Python}},
+  year =      {2001--},
+  url = "http://www.scipy.org/",
+  note = {[Online; accessed <today>]}
+}
+
+
+@article{matplotlib:2007,
+author = {John D. Hunter},
+title = {Matplotlib: A 2D Graphics Environment},
+year = {2007},
+journal= {Computing in Science \& Engineering},
+volume = {9}
+pages = {90-95},
+doi = {DOI:10.1109/MCSE.2007.55}
+}
+
+@article{numpy:2011,
+author = {St\'{e}fan van der Walt, S. Chris Colbert and Ga\"{e}l Varoquaux},
+title = {"The NumPy Array: A Structure for Efficient Numerical Computation},
+journal = {Computing in Science \& Engineering},
+volume = {13},
+pages = {22-30},
+year = {2011},
+doi = {DOI:10.1109/MCSE.2011.37}
+}
+
+
+
+
+
