### 2017-08-29 e0e5897
Create CONTRIBUTING.rst
