+++ CONTRIBUTING.rst
@@ -1,0 +1,33 @@
+==================================
+Spectrum contributing guidelines
+==================================
+
+Issues
+==========
+
+You have an issue ? You found a bug ? Please submit an issue with 
+
+- a description of the problem
+- the version of Sequana used
+- the error message if any
+
+
+pull requests
+===================
+
+Pull requests are always welcome, and the Spectrum community appreciates
+any help you may give.
+
+When submitting a pull request, we ask you to check the following:
+
+1. **Unit tests**, **documentation**, and **code style** are in order. 
+   
+   It's also OK to submit work in progress if you're unsure of what
+   this exactly means, in which case you'll likely be asked to make
+   some further changes.
+
+2. The contributed code will be **licensed under spectrum's license** (BSD3),
+   If you did not write the code yourself, you ensure the existing
+   license is compatible and include the license information in the
+   contributed files, or obtain a permission from the original
+   author to relicense the contributed code.
