### 2017-08-29 #348 @cokelaer
@eteq @arfon Here are some answers to the checklist.Thanks again for the suggestions/comments

-  no clear statement of need: I have edited the README.rst file adding

   "The targetted audience is diverse. Although the use of power spectrum of a signal is fundamental in electrical engineering (e.g. radio communications, radar), it has a wide range of applications from cosmology (e.g., detection of gravitational waves in 2016), to music (pattern detection) or biology (mass spectroscopy)."

This is also reflected in the readthedocs documentation and paper.md for JOSS.

- examples are not "real world" examples. Indeed,  mostly used a complex (math) example  and in some cases a sinus signal (one frequency). Although note that in the quick start example, this is a real-case example with the detection of a mono frequency signal. To improve this aspect of the documentation, I have added a gallery that will be completed little by little (http://pyspectrum.readthedocs.io/en/latest/auto_examples/index.html). For now, I have put 3 examples. One is a real case example (time-frequency detection of a dolphin "song") and the two others are still a bit theoretical (addition of two sinus signals).

- No clear community guidelines: I have added a contributing.rst file

- only one author on the paper:  I have added an  AUTHORS.rst with the link to the github account of the contributors. As for the JOSS paper, I have only added my name. This is a tricky question. I believe that most of the original work back from 2011 was done by myself and then small contributions were done by a few developers. I'm not sure how to handle this. Shall I add someone for fixing a typo in a comment for instance ? Where to you put the threshold ? Also as you said, what about personnal communication and bug report by email ? So, for now, I decided to go for a single author in the joss document while adding contributors in a dedicated file. I am open to discussion of course, but I think it is fair as it is. I could also contact the main contributors to ask their opinion.

- "While various packages... citations...": I have added numpy/matplotlib/scipy in the paper.bib and added a short sentence at the end to explain what is used as dependencies.

- "I found spectrum to be a rather confusing name": *spectrum* is quite generic indeed but used by a few people so I will stick to it.
