### 2017-08-29 #348 @cokelaer
Note also that since last two weeks, I have cleaned the code significantly, fixing a few bugs and adding tools such as spectrogram. These updates are available in the release 0.7.1 so we may want to update the version 0.6.2 to 0.7.1 in the JOSS review. thanks
