### 2017-09-20 #348 @cokelaer
@arfon @eteq
Yes, I had updated the code and documentation. please go ahead. thanks
