### 2017-10-06 #348 @eteq
@cokelaer - thanks for your work here.  You've addressed all my concerns - we're almost there, but I want to push just a little on one of these. (But want to make it clear at the outset that I am *not* taking a hard-line stance and am open to discussion up to and including accepting everything as-is):

>  As for the JOSS paper, I have only added my name. This is a tricky question. I believe that most of the original work back from 2011 was done by myself and then small contributions were done by a few developers. I'm not sure how to handle this. Shall I add someone for fixing a typo in a comment for instance ? Where to you put the threshold ? Also as you said, what about personnal communication and bug report by email ? So, for now, I decided to go for a single author in the joss document while adding contributors in a dedicated file. I am open to discussion of course, but I think it is fair as it is. I could also contact the main contributors to ask their opinion.

My usual stance inside my field is that anyone who has contributed code *or* significant non-code effort should be a co-author of a package.  Email communication/bug reports in general I wouldn't count, *unless* they involve substantive effort. I.e., someone emailed you a patch because they're not git-fluent, but if they had been, they would have issued a PR... Nut I would not count someone who emailed you once "hey maybe you should add feature X because I like it".   And it seems to me that the "package authors" = "JOSS paper authors", given that if they deserve credit for the software, they should get some credit through a peer review service for the software (roughly speaking, that's JOSS).

All that said, I recognize all this is driven quite a bit by the social norms of my discipline: in astronomy we tend to have long author lists but recognize that the first (or at least a small subset) of the authors gets most of the credit for doing the work.  In that scheme, having lots of extra authors isn't a problem and it's sort of rude *not* to include them.  In other fields I know this is not the case.  So if you think this is *not* the norm for your field, I'll accept that, but I want to be sure you've thought about it in those terms.  And if you're not sure on that basis, I think the "ask the main contributors" is a good solution (although with a timeline - e.g. if they don't respond within a week or something, they forfeit the authorship option).
