### 2017-10-06 #348 @arfon
@cokelaer @eteq - Facing a similar issue, `@adrn` [opened an issue](https://github.com/openjournals/joss-reviews/issues/388#issuecomment-333407285) in his repository to check with other contributors https://github.com/adrn/gala/issues/100

> All that said, I recognize all this is driven quite a bit by the social norms of my discipline: in astronomy we tend to have long author lists but recognize that the first (or at least a small subset) of the authors gets most of the credit for doing the work. In that scheme, having lots of extra authors isn't a problem and it's sort of rude not to include them. In other fields I know this is not the case. So if you think this is not the norm for your field, I'll accept that, but I want to be sure you've thought about it in those terms. And if you're not sure on that basis, I think the "ask the main contributors" is a good solution (although with a timeline - e.g. if they don't respond within a week or something, they forfeit the authorship option).

To be clear, JOSS is not here to police authorship and I think @eteq has done a good job of articulating his position here. @cokelaer - if you're interested in doing something similar to `@adrn` in https://github.com/adrn/gala/issues/100 then we can wait for a week or so before accepting, otherwise, we can move forward with accepting this submission. Your choice 😸
