### 2017-10-06 cokelaer/spectrum#41 @cokelaer
Authorship on JOSS paper
@juhasch @anielsen001 @carlkl
I've submitted a paper to [JOSS](https://github.com/openjournals/) describing the Spectrum package. It's currently under review, and you can read the review and response at the [link.](https://github.com/openjournals/joss-reviews/issues/348).

Please let me know by 16 Oct if you'd like to be listed as co-authors on the paper, in which case please send me your current affiliation and ORCID.
