### 2017-10-06 #348 @cokelaer
@arfon @eteq
I think that it is a valid point and agree with the idea of adding authors with significant contributions. The solution proposed . I have open an issue accordingly https://github.com/cokelaer/spectrum/issues/41
