### 2017-10-07 #348 @eteq
Fantastic! Once your deadline from cokelaer/spectrum#41 has passed, you can consider this as having been accepted, @arfon and @cokelaer
