### 2017-10-09 cokelaer/spectrum#41 @juhasch
Hi @cokelaer,
not sure if my not so large contribution qualifies me as co-author. If you think that's appropriate:
http://orcid.org/0000-0002-9457-1220
