### 2017-10-10 8eaa44d
added Taylor window
