### 2017-10-10 8eb305d
fixed normalization in Taylor window
