### 2017-10-11 cokelaer/spectrum#41 @cokelaer
@juhasch we'll add you to the author list. thanks for your contributions
