+++ src/spectrum/window.py
@@ -1186,7 +1186,7 @@ def window_taylor(N, nbar=4, sll=-30):
         return numer/denom
     Fm = array([calc_Fm(m) for m in ma])
     def W(n):
-        return 2*Fm @ cos(2*pi*ma*(n-N/2 + 1/2)/N) + 1
+        return 2*np.sum(Fm * cos(2*pi*ma*(n-N/2 + 1/2)/N)) + 1
     w = array([W(n) for n in range(N)])
     # normalize (Note that this is not described in the original text)
     scale = W((N-1)/2)
