### 2017-10-11 cokelaer/spectrum#41 @stringhamc
Hi @cokelaer, I appreciate the offer. At this point there is no need to add me as an author as my only contribution is implementing a previously documented function. But I'll check out the paper. Thanks again!
