### 2017-10-16 #348 @arfon
> Fantastic! Once your deadline from cokelaer/spectrum#41 has passed, you can consider me as signed off on this as accepted, @arfon and @cokelaer

@cokelaer - are we good to proceed?
