### 2017-10-18 cokelaer/spectrum#41 @cokelaer
@juhasch and @stringhamc  so I will proceed with the finalisation of the paper.
Just to be fair with open source contributions, your commits were useful and if you want you can just update this file in the spectrum repo:

https://github.com/cokelaer/spectrum/blob/master/doc/joss/paper.md

adding your name/affiliation (if relevant) and ORCID.  If you could do so before the end of the week, that would be great so that I can finalise the JOSS submission.
best
