### 2017-10-18 #348 @cokelaer
I've asked two contributors to update the metadata related to authorship. Let us wait a couple of days. thanks
