### 2017-10-20 cokelaer/spectrum#41 @cokelaer
FYI @juhasch @stringhamc reading back your messages in more details, I have included Juergen only in the JOSS meta data information in the author list.
