### 2017-10-23 #348 @cokelaer
@arfon  @eteq

I have added one additional author in the paper following public discussion at https://github.com/cokelaer/spectrum/issues/41. I believe we are ready to proceed.
