### 2017-10-24 7c90628
Fixing paper metadata
