+++ doc/joss/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'Spectrum': Spectral Analysis in Python
+title: "'Spectrum': Spectral Analysis in Python"
 tags:
   - spectral analysis
   - periodogram
@@ -16,11 +16,10 @@ authors:
  - name: Juergen Hasch
    orcid: 0000-0002-9457-1220
    affiliation: 2
- affiliations:
- - name: Institut Pasteur - Bioinformatics and Biostatistics Hub - C3BI, USR 3756 IP
-CNRS - Paris, France
+affiliations:
+ - name: Institut Pasteur - Bioinformatics and Biostatistics Hub - C3BI, USR 3756 IP CNRS - Paris, France
    index: 1
- - name:  Robert Bosch GmbH: Renningen, Baden-Württemberg, Germany
+ - name: "Robert Bosch GmbH: Renningen, Baden-Württemberg, Germany"
    index: 2
 date: 2 August 2017
 bibliography: paper.bib
@@ -28,12 +27,12 @@ bibliography: paper.bib
 
 # Summary
 
-**Spectrum** is a Python library that includes tools to estimate Power Spectral Densities. Although the use of 
+**Spectrum** is a Python library that includes tools to estimate Power Spectral Densities. Although the use of
 power spectrum of a signal is fundamental in electrical engineering (e.g. radio communications, radar), it has
-a wide range of applications from cosmology (e.g., detection of gravitational waves in 2016), to music 
+a wide range of applications from cosmology (e.g., detection of gravitational waves in 2016), to music
 (pattern detection) or biology (mass spectroscopy).
 
-Methods available are based on Fourier transform, parametric methods or eigenvalues analysis. Although standard methods such as 
+Methods available are based on Fourier transform, parametric methods or eigenvalues analysis. Although standard methods such as
 periodogram are available, less common methods (e.g. multitapering) are also implemented:
 
 - The Fourier methods are based upon correlogram, periodogram and Welch estimates. Standard tapering windows (Hann, Hamming, Blackman) and more exotic ones are available (DPSS, Taylor, ...)[harris:1978],[welch:1967],[marple:1987].
