### 2017-10-24 #348 @arfon
@cokelaer - thanks for the update. Could you please merge this PR with some small fixes https://github.com/cokelaer/spectrum/pull/43 ? Also, please make sure you're citing the entries in your `paper.bib` file correctly (looks like they need a `@` before the bibtex key) otherwise they won't be compiled into the final PDF. You can read how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html)
