### 2017-10-25 83938b6
Update paper.md
