### 2017-10-26 ba2d842
update metadata (new version) and changelog
