### 2017-10-26 #348 @cokelaer
@arfon, zenodo was already setup. I just have tag a new release https://github.com/cokelaer/spectrum/releases that is now also on zenodo with the following DOI: https://doi.org/10.5281/zenodo.1037268
